update db schema and get receipt

+++ migrations/2023-08-15-135134_create_coinbases/up.sql
@@ -2,11 +2,11 @@
 create table
   coinbases (
     id character varying(255) not null,
-    height integer null,
-    da_height integer null,
+    height BIGINT null,
+    da_height BIGINT null,
     block_hash character varying(255) not null,
     amount character varying(255) null,
     coinbase character varying(255) null,
-    timestamp integer null,
+    timestamp BIGINT null,
     constraint coinbases_pkey primary key (id)
   ) tablespace pg_default;
