+++ migrations/2023-08-15-135146_create_accounts/up.sql
@@ -2,6 +2,6 @@
 create table
   accounts (
     id character varying(255) not null,
-    timestamp integer null,
+    timestamp BIGINT null,
     constraint accounts_pkey primary key (id)
   ) tablespace pg_default;
