+++ migrations/2023-08-15-135213_create_assets/up.sql
@@ -2,6 +2,6 @@
 create table
   assets (
     id character varying(255) not null,
-    timestamp integer null,
+    timestamp BIGINT null,
     constraint assets_pkey primary key (id)
   ) tablespace pg_default;
