+++ migrations/2023-08-15-135243_create_nfts/up.sql
@@ -2,6 +2,6 @@
 create table
   nfts (
     id character varying(255) not null,
-    timestamp integer null,
+    timestamp BIGINT null,
     constraint nfts_pkey primary key (id)
   ) tablespace pg_default;
